Поправлена файловая структура, поправлено меню юзера в хедере
