--- client/src/UI/Icon/index.ts
@@ -1,1 +1,0 @@
-export { IconTemplate as Icon } from './IconTemplate'
