+++ client/src/components/App/AppBehavior.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Provider } from 'react-redux'
 import { configureStore } from '../../redux/authToolkitRedux'
