+++ client/src/components/App/AppTemplate.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Route, Switch, BrowserRouter as Router, Redirect } from 'react-router-dom'
