+++ client/src/components/AppLayout/AppLayout.scss
@@ -53,6 +53,7 @@
   text-overflow: ellipsis;
   margin-bottom: 0;
   margin-right: 10px;
+  color: #1890ff;
 }
 
 .app {
