+++ client/src/components/AppLayout/AppLayout.tsx
@@ -1,28 +1,27 @@
-import React from 'react'
-import { useSelector } from 'react-redux'
-import { Layout, Menu, Avatar } from 'antd'
+import { useState, FC } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { Layout, Menu } from 'antd'
 import { UserOutlined } from '@ant-design/icons'
 
-import { User } from '../User'
+import { User } from './User'
 import { routes } from './appRoutes'
 import { MENU_SIDEBAR_ITEMS } from '../../constants/constants'
 import './AppLayout.scss'
 import { RootState } from 'redux/authToolkitRedux/StoreSlices'
 import { Link, useLocation } from 'react-router-dom'
+import { fetchLogout } from 'pages/Authorization/authSlice'
 
 const { Header, Sider, Content } = Layout
 
-interface IProps {
-  collapsed: boolean
-  setCollapsed: any
-  onHandleLogout(): void
-}
+interface IProps {}
 
-export const AppLayoutTemplate = ({
-  collapsed,
-  setCollapsed,
-  onHandleLogout,
-}: IProps): JSX.Element => {
+export const AppLayout: FC<IProps> = () => {
+  const dispatch = useDispatch()
+  const [collapsed, setCollapsed] = useState(false)
+
+  const onHandleLogout = () => {
+    dispatch(fetchLogout())
+  }
   const location = useLocation()
   const activeItem = MENU_SIDEBAR_ITEMS.find((item) => location.pathname.includes(item.key))
   const activeItems = activeItem ? [activeItem.key] : undefined
@@ -56,10 +55,9 @@ export const AppLayoutTemplate = ({
       </Sider>
       <Layout className={`${ROOT_CLASS}-wrapper`}>
         <Header className={`${ROOT_CLASS}-wrapper__header`}>
-          <div className={`${ROOT_CLASS}-wrapper__header-user`} onClick={onHandleLogout}>
+          <div className={`${ROOT_CLASS}-wrapper__header-user`}>
             {/* TODO поменять на ник */}
-            <User user={{ name: email }} />
-            <Avatar icon={<UserOutlined />} />
+            <User user={{ name: email }} onLogout={onHandleLogout} />
           </div>
         </Header>
         <Content
