+++ client/src/components/Feed/Feed.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Spin } from 'antd'
 import InfiniteScroll from 'react-infinite-scroll-component'
