+++ client/src/pages/AddMeme/AddMeme.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { serialize } from 'object-to-formdata'
