+++ client/src/pages/Profile/ProfileTemplate.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Avatar } from 'antd'
 import { UserOutlined } from '@ant-design/icons'
