+++ client/src/pages/Statistics/StatisticsTemplate.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export const StatisticsTemplate = () => {
   return <div>Статистика</div>
 }
